+++ packages/indexer/src/api/endpoints/tokens/get-user-tokens/v7.ts
@@ -28,8 +28,8 @@ import {
 import { config } from "@/config/index";
 import { CollectionSets } from "@/models/collection-sets";
 import { Sources } from "@/models/sources";
-import { isOrderNativeOffChainCancellable } from "@/orderbook/orders/utils";
 import { Assets, ImageSize } from "@/utils/assets";
+import { isOrderNativeOffChainCancellable } from "@/utils/offchain-cancel";
 
 const version = "v7";
