+++ packages/indexer/src/utils/offchain-cancel/index.ts
@@ -1,4 +1,6 @@
+import { AddressZero } from "@ethersproject/constants";
 import { Wallet } from "@ethersproject/wallet";
+import * as Sdk from "@reservoir0x/sdk";
 
 import { idb, pgp } from "@/common/db";
 import { config } from "@/config/index";
@@ -59,4 +61,19 @@ export const saveOffChainCancellations = async (orderIds: string[]) => {
       },
     }))
   );
+};
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export const isOrderNativeOffChainCancellable = (rawData?: any) => {
+  // Seaport
+  if (rawData?.zone) {
+    return rawData.zone === Sdk.SeaportBase.Addresses.ReservoirCancellationZone[config.chainId];
+  }
+
+  // Payment Processor
+  if (rawData?.cosigner) {
+    return rawData.cosigner !== AddressZero;
+  }
+
+  return false;
 };
